# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para conversion de PDF, IIIF Presentation v2 (con v3 compatible), IIIF Image API, administracion y migracion. Las rutas recomendadas usan /api/v1 y los endpoints legacy se mantienen por compatibilidad temporal.",
        "title": "IIIF PDF Server API",
        "contact": {},
        "version": "2.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/iiif/{id}/manifest": {
            "get": {
                "description": "Devuelve un manifiesto v2 con secuencia, canvases y rangos jerarquicos derivados de los marcadores PDF. El parametro pages permite seleccionar paginas y poda rangos vacios.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifestV2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/documents": {
            "get": {
                "description": "Lista documentos con filtros opcionales por proyecto y tenant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentos"
                ],
                "summary": "Listar documentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proyecto",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PDFDocument"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/upload": {
            "post": {
                "description": "Recibe archivo PDF y crea documento con paginas convertidas.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documentos"
                ],
                "summary": "Subir y convertir PDF",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Archivo PDF",
                        "name": "pdf",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Proyecto",
                        "name": "project",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Tenant",
                        "name": "tenant",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 1241,
                        "description": "Ancho maximo de imagen",
                        "name": "max_width",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 1754,
                        "description": "Alto maximo de imagen",
                        "name": "max_height",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 150,
                        "description": "Resolucion de render",
                        "name": "dpi",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "jpg",
                        "description": "Formato de salida (jpg o png)",
                        "name": "format",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 85,
                        "description": "Calidad JPEG",
                        "name": "quality",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PDFDocument"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/documents/{id}/ocr": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Obtener estado OCR de un documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.OCRDocumentSummary"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/documents/{id}/ocr/pages/{page}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OCR"
                ],
                "summary": "Obtener OCR y bounding boxes de una página",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.OCRPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/iiif/{id}/manifest": {
            "get": {
                "description": "Devuelve un manifiesto v2 con secuencia, canvases y rangos jerarquicos derivados de los marcadores PDF. El parametro pages permite seleccionar paginas y poda rangos vacios.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifestV2"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/iiif/{id}/manifest/v3": {
            "get": {
                "description": "Endpoint de compatibilidad para clientes IIIF Presentation v3.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener manifiesto IIIF Presentation API v3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del documento",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Paginas o rangos, por ejemplo 1-3,5",
                        "name": "pages",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFManifest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "Crea cookie de sesion para rutas administrativas.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesion de dashboard",
                "parameters": [
                    {
                        "description": "Credenciales",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.sessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/iiif/3/{identifier}/info.json": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IIIF"
                ],
                "summary": "Obtener info IIIF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador IIIF",
                        "name": "identifier",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.IIIFImageInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "detalle interno"
                },
                "error": {
                    "type": "string",
                    "example": "login requerido"
                }
            }
        },
        "handlers.loginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "CAMBIAR_PASSWORD"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "models.IIIFManifest": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFCanvas"
                    }
                },
                "label": {},
                "type": {
                    "type": "string"
                }
            }
        },
        "models.IIIFManifestV2": {
            "type": "object",
            "properties": {
                "@context": {
                    "type": "string"
                },
                "@id": {
                    "type": "string"
                },
                "@type": {
                    "type": "string"
                },
                "attribution": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "license": {
                    "type": "string"
                },
                "metadata": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFMetadataV2"
                    }
                },
                "sequences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFSequenceV2"
                    }
                },
                "structures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.IIIFRangeV2"
                    }
                },
                "thumbnail": {
                    "$ref": "#/definitions/models.IIIFResourceV2"
                }
            }
        },
...
        "models.PDFDocument": {
            "type": "object",
            "properties": {
                "conversionDpi": {
                    "type": "integer"
                },
                "conversionFormat": {
                    "type": "string"
                },
                "conversionHeight": {
                    "type": "integer"
                },
                "conversionQuality": {
                    "type": "integer"
                },
                "conversionWidth": {
                    "type": "integer"
                },
                "convertedPages": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "manifestUrl": {
                    "type": "string"
                },
                "migratedFromLocal": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "outline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PDFOutlineItem"
                    }
                },
                "projectKey": {
                    "type": "string"
                },
                "status": {
                    "description": "processing, completed, error",
                    "type": "string"
                },
                "tenantKey": {
                    "type": "string"
                },
                "thumbnailUrl": {
                    "type": "string"
                },
                "totalPages": {
                    "type": "integer"
                },
                "uploadDate": {
                    "type": "string"
                }
            }
        },
...
        "services.OCRDocumentSummary": {
            "type": "object",
            "properties": {
                "active_generation": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "failed_pages": {
                    "type": "integer"
                },
                "indexed_pages": {
                    "type": "integer"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "project_key": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tenant_key": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "services.OCRPage": {
            "type": "object",
            "properties": {
                "canvas_v2": {
                    "type": "string"
                },
                "canvas_v3": {
                    "type": "string"
                },
                "confidence": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "engine": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "generation": {
                    "type": "string"
                },
                "geometry_status": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "native_text": {
                    "type": "string"
                },
                "ocr_text": {
                    "type": "string"
                },
                "page_number": {
                    "type": "integer"
                },
                "schema_version": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "words": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.OCRWord"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "project_iiif_session",
            "in": "cookie"
        }
    }
}